Move all lsp services to own lsp directory
