+++ libs/language-server/src/lib/lsp/index.ts
@@ -2,4 +2,6 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-export * from './runtime-parameter-provider';
+export * from './jayvee-completion-provider';
+export * from './jayvee-formatter';
+export * from './jayvee-hover-provider';
