+++ libs/language-server/src/lib/lsp/jayvee-completion-provider.ts
