+++ libs/language-server/src/lib/lsp/jayvee-formatter.ts
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2023 Friedrich-Alexander-Universitat Erlangen-Nurnberg
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
 import { type AstNode, type CstNode, isCompositeCstNode } from 'langium';
 import {
   AbstractFormatter,
