+++ libs/language-server/src/lib/lsp/jayvee-hover-provider.ts
